feat: ignore action, ignore test migration
