+++ src/main/kotlin/com/jaoafa/vcspeaker/tts/narrators/Narrator.kt
@@ -3,8 +3,8 @@ package com.jaoafa.vcspeaker.tts.narrators
 import com.jaoafa.vcspeaker.VCSpeaker
 import com.jaoafa.vcspeaker.database.actions.GuildAction.getVoice
 import com.jaoafa.vcspeaker.database.actions.GuildAction.getVoiceTextChannelOrNull
+import com.jaoafa.vcspeaker.database.actions.IgnoreAction
 import com.jaoafa.vcspeaker.database.actions.UserAction
-import com.jaoafa.vcspeaker.features.Ignore.shouldBeIgnored
 import com.jaoafa.vcspeaker.reload.state.UseState
 import com.jaoafa.vcspeaker.tools.discord.DiscordExtensions.addReactionSafe
 import com.jaoafa.vcspeaker.tools.discord.DiscordExtensions.deleteOwnReactionSafe
@@ -102,7 +102,7 @@ class Narrator @OptIn(KordVoice::class) constructor(
         guild: Guild,
         actor: SpeechActor
     ) {
-        if (shouldBeIgnored(text, guildId)) return
+        if (IgnoreAction.shouldBeIgnored(text, guildId)) return
 
         val sounds = soundRegex.findAll(text).mapNotNull {
             val id = it.groupValues[1].toLongOrNull() ?: return@mapNotNull null
