+++ src/main/kotlin/com/jaoafa/vcspeaker/tts/processors/IgnoreBeforeReplaceProcessor.kt
@@ -1,6 +1,6 @@
 package com.jaoafa.vcspeaker.tts.processors
 
-import com.jaoafa.vcspeaker.features.Ignore.shouldBeIgnored
+import com.jaoafa.vcspeaker.database.actions.IgnoreAction
 import com.jaoafa.vcspeaker.tts.Voice
 import dev.kord.core.entity.Message
 
@@ -9,7 +9,7 @@ class IgnoreBeforeReplaceProcessor : BaseProcessor() {
 
     override suspend fun process(message: Message?, content: String, voice: Voice): Pair<String, Voice> {
         val guildId = message?.getGuild()?.id ?: return content to voice
-        if (shouldBeIgnored(content, guildId)) cancel()
+        if (IgnoreAction.shouldBeIgnored(content, guildId)) cancel()
         return content to voice
     }
 }
