+++ src/test/kotlin/processors/CharLimitProcessorTest.kt
@@ -1,8 +1,8 @@
 package processors
 
 import com.jaoafa.vcspeaker.tts.Voice
-import com.jaoafa.vcspeaker.tts.providers.voicetext.Speaker
 import com.jaoafa.vcspeaker.tts.processors.CharLimitProcessor
+import com.jaoafa.vcspeaker.tts.providers.voicetext.Speaker
 import dev.kord.core.entity.Message
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
