+++ src/test/kotlin/processors/InlineVoiceProcessorTest.kt
@@ -2,9 +2,9 @@ package processors
 
 import com.jaoafa.vcspeaker.tts.EmotionData
 import com.jaoafa.vcspeaker.tts.Voice
+import com.jaoafa.vcspeaker.tts.processors.InlineVoiceProcessor
 import com.jaoafa.vcspeaker.tts.providers.voicetext.Emotion
 import com.jaoafa.vcspeaker.tts.providers.voicetext.Speaker
-import com.jaoafa.vcspeaker.tts.processors.InlineVoiceProcessor
 import dev.kord.core.entity.Message
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
