+++ src/test/kotlin/processors/ReplyProcessorTest.kt
@@ -1,8 +1,8 @@
 package processors
 
 import com.jaoafa.vcspeaker.tts.Voice
-import com.jaoafa.vcspeaker.tts.providers.voicetext.Speaker
 import com.jaoafa.vcspeaker.tts.processors.ReplyProcessor
+import com.jaoafa.vcspeaker.tts.providers.voicetext.Speaker
 import dev.kord.common.entity.MessageType
 import dev.kord.core.entity.Message
 import dev.kord.core.entity.effectiveName
