+++ src/test/kotlin/providers/SpeechProviderTest.kt
@@ -1,16 +1,19 @@
 package providers
 
-import com.jaoafa.vcspeaker.tts.providers.*
+import com.jaoafa.vcspeaker.tts.Voice
+import com.jaoafa.vcspeaker.tts.providers.ProviderContext
+import com.jaoafa.vcspeaker.tts.providers.getProvider
+import com.jaoafa.vcspeaker.tts.providers.hashMd5
+import com.jaoafa.vcspeaker.tts.providers.providerOf
 import com.jaoafa.vcspeaker.tts.providers.soundmoji.SoundmojiContext
 import com.jaoafa.vcspeaker.tts.providers.soundmoji.SoundmojiProvider
+import com.jaoafa.vcspeaker.tts.providers.voicetext.Speaker
 import com.jaoafa.vcspeaker.tts.providers.voicetext.VoiceTextContext
 import com.jaoafa.vcspeaker.tts.providers.voicetext.VoiceTextProvider
-import com.jaoafa.vcspeaker.tts.Voice
-import com.jaoafa.vcspeaker.tts.providers.voicetext.Speaker
 import dev.kord.common.entity.Snowflake
+import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
-import io.kotest.assertions.throwables.shouldThrow
 
 class SpeechProviderTest : FunSpec({
     test("hashMd5 should compute correct MD5 hash") {
