+++ src/test/kotlin/stores/AliasStoreSerializationTest.kt
@@ -6,13 +6,7 @@ import com.jaoafa.vcspeaker.stores.TypedStore
 import dev.kord.common.entity.Snowflake
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonArray
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.buildJsonObject
-import kotlinx.serialization.json.jsonArray
-import kotlinx.serialization.json.jsonObject
+import kotlinx.serialization.json.*
 
 class AliasStoreSerializationTest : FunSpec({
     test("AliasData ignores unknown keys during deserialization") {
