+++ src/test/kotlin/tools/VisionApiTest.kt
@@ -2,8 +2,8 @@ package tools
 
 import com.jaoafa.vcspeaker.tools.VisionApi
 import com.jaoafa.vcspeaker.tools.VisionApi.VisionApiUnsupportedMimeTypeException
-import io.kotest.core.spec.style.FunSpec
 import io.kotest.assertions.throwables.shouldThrow
+import io.kotest.core.spec.style.FunSpec
 
 class VisionApiTest : FunSpec({
     test("getTextAnnotations throws for unsupported MIME type") {
